Move API "public" headers to their own directory.
As suggested by Hugh Sorby (@hsorby).  This has a few benefits, for
examp.e, it helps validate that the set of headers is correct, since
the unit tests can only see the public headers.  Previously, headers
that were not intended to be part of the public API might also be in
the `src` directory.

(Obviously at this stage we only have one header file in the
set of public headers, but the plural is used since this will change
soon.)
