+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(HEADER_FILES 
-  ${PROJECT_SOURCE_DIR}/src/version.h
+  ${PROJECT_SOURCE_DIR}/src/libcellml/version.h
 )
 
 configure_file (
@@ -8,6 +8,7 @@ configure_file (
 )
 
 include_directories(
+  "${PROJECT_BINARY_DIR}/src/libcellml"
   "${PROJECT_BINARY_DIR}/src"
 )
